requirements.txt
streamlit==1.47.0
pandas==2.3.1
langchain==0.3.27
langchain-core==0.3.72
langchain-community==0.3.27
langchain-openai==0.3.28
langchain-qdrant==0.2.0
qdrant-client==1.15.0
python-dotenv==1.1.1
lark==1.2.2
